requirements_cuda.txt
### Pytorch ###

--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1

### Pytorch Geometric ###
torch_geometric

-f https://data.pyg.org/whl/torch-2.7.0+cu128.html
pyg_lib
torch_scatter
torch_sparse
torch_cluster
torch_spline_conv


### Pytorch Extras ###

torchtyping
torchmetrics
vector-quantize-pytorch

### Hydra ###

hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
hydra-ray-launcher==1.2.1

### Hugging Face ###

accelerate
diffusers
timm
transformers
diffusers

### AI utils ###

einops
einx
local-attention
ocnn
taylor-series-linear-attention
x_transformers

### Python utils ###

beartype
black
blobfile
h5py
imageio
matplotlib
notebook
numpy
opencv-python
pandas
plyfile
pre-commit
pudb
pykdtree
# pyrender # requires numpy < 2.0
git+https://github.com/hilookas/pyrender@master#egg=pyrender
PyOpenGL==3.1.9
PyOpenGL-accelerate
pytest
scikit-image
scikit-learn
scipy
setuptools
transforms3d
yacs

### Compile utils ###

cmake
cython
ninja

### Log & OS ###

coloredlogs
rootutils
rich
sh
termcolor
tqdm
tqdm_joblib
wget

### Mesh Processing ###

mesh2sdf
open3d>=0.19.0
PyMCubes
trimesh

### Loggers ###

tensorboard
tensorboard-plugin-profile
tensorboardX
# wandb
# neptune-client
# mlflow
# comet-ml
aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

### Blender ###
# bpy
# mathutils

### Benchmark ###
libigl
sentence-transformers
lpips